Implement wallet backups for SQLite
