+++ crates/cdk-cli/src/main.rs
@@ -102,19 +102,11 @@ async fn main() -> Result<()> {
 
     fs::create_dir_all(&work_dir)?;
 
+    let backups_to_keep = 3;
     let localstore: Arc<dyn WalletDatabase<Err = cdk_database::Error> + Send + Sync> =
         match args.engine.as_str() {
-            "sqlite" => {
-                let sql_path = work_dir.join("cdk-cli.sqlite");
-                let sql = WalletSqliteDatabase::new(&sql_path).await?;
-
-                Arc::new(sql)
-            }
-            "redb" => {
-                let redb_path = work_dir.join("cdk-cli.redb");
-
-                Arc::new(WalletRedbDatabase::new(&redb_path)?)
-            }
+            "sqlite" => Arc::new(WalletSqliteDatabase::new(&work_dir, backups_to_keep).await?),
+            "redb" => Arc::new(WalletRedbDatabase::new(&work_dir)?),
             _ => bail!("Unknown DB engine"),
         };
