+++ crates/cdk-sqlite/src/mint/mod.rs
@@ -1,9 +1,8 @@
 //! SQLite Mint
 
 use std::collections::HashMap;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::str::FromStr;
-use std::time::Duration;
 
 use async_trait::async_trait;
 use bitcoin::bip32::DerivationPath;
@@ -19,12 +18,13 @@ use cdk_common::{
     State,
 };
 use lightning_invoice::Bolt11Invoice;
-use sqlx::sqlite::{SqliteConnectOptions, SqlitePool, SqlitePoolOptions, SqliteRow};
-use sqlx::{Pool, Row, Sqlite};
+use sqlx::sqlite::{SqlitePool, SqliteRow};
+use sqlx::Row;
 use uuid::fmt::Hyphenated;
 use uuid::Uuid;
 
 use crate::error::Error;
+use crate::{backup, connect_to_db};
 
 /// Mint SQLite Database
 #[derive(Debug, Clone)]
@@ -43,86 +43,15 @@ impl MintSqliteDatabase {
         let db_file_path = work_dir.join("cdk-mintd.sqlite");
 
         if db_file_path.exists() {
-            Self::backup(work_dir, &db_file_path, backups_to_keep).await?;
+            backup(work_dir, &db_file_path, backups_to_keep).await?;
         }
 
-        let pool = Self::connect_to_db(&db_file_path).await?;
+        let pool = connect_to_db(&db_file_path).await?;
         let db = Self { pool: pool.clone() };
 
         sqlx::migrate!("./src/mint/migrations").run(&pool).await?;
 
         Ok(db)
-    }
-
-    async fn backup(
-        work_dir: &Path,
-        db_file_path: &Path,
-        backups_to_keep: u8,
-    ) -> Result<(), Error> {
-        if let Some(new_backup_file) = database::prepare_backup(work_dir, "sqlite", backups_to_keep)
-            .map_err(Error::DbBackup)?
-        {
-            Self::create_backup(db_file_path, new_backup_file)
-                .await
-                .map_err(Error::DbBackup)?;
-        }
-
-        Ok(())
-    }
-
-    async fn connect_to_db(db_file_path: &Path) -> Result<Pool<Sqlite>, Error> {
-        let db_file_path_str = db_file_path.to_str().ok_or(Error::InvalidDbPath)?;
-        let db_options = SqliteConnectOptions::from_str(db_file_path_str)?
-            .busy_timeout(Duration::from_secs(5))
-            .read_only(false)
-            .create_if_missing(true)
-            .auto_vacuum(sqlx::sqlite::SqliteAutoVacuum::Full);
-
-        let pool = SqlitePoolOptions::new()
-            .max_connections(1)
-            .connect_with(db_options)
-            .await?;
-
-        Ok(pool)
-    }
-
-    async fn create_backup(
-        db_file_path: &Path,
-        backup_file_path: PathBuf,
-    ) -> Result<(), database::Error> {
-        let pool = Self::connect_to_db(db_file_path).await?;
-
-        let backup_path_str = backup_file_path
-            .to_str()
-            .ok_or_else(|| database::Error::Database("Invalid backup path".into()))?;
-
-        // Create a backup connection with the destination path
-        let backup_options = SqliteConnectOptions::from_str(backup_path_str)
-            .map_err(|e| {
-                database::Error::Database(format!("Failed to create backup options: {e}").into())
-            })?
-            .create_if_missing(true);
-
-        let backup_pool = SqlitePoolOptions::new()
-            .max_connections(1)
-            .connect_with(backup_options)
-            .await
-            .map_err(|e| {
-                database::Error::Database(format!("Failed to create backup connection: {e}").into())
-            })?;
-
-        // Execute backup
-        sqlx::query("VACUUM INTO ?")
-            .bind(backup_path_str)
-            .execute(&pool)
-            .await
-            .map_err(|e| {
-                database::Error::Database(format!("Failed to create backup: {e}").into())
-            })?;
-
-        backup_pool.close().await;
-
-        Ok(())
     }
 }
